requirements.txt
pandas
matplotlib
seaborn
gradio
openai
python-dotenv